New 'dup' verb. Used heavily by rq_router

+++ code/test/setup_test_queues.sh
@@ -107,11 +107,25 @@ if [ "$?" -ne "0" ]; then
   exit 1
 fi
 
+echo "Creating the test nop queue..."
+curl -0 http://127.0.0.1:${rq_port}/new_queue -sL -F queue[name]=test_nop -F queue[script]=./code/test/test_nop.sh -F queue[num_workers]=1 -o _install_test_nop.txt
+if [ "$?" -ne "0" ]; then
+  echo "Sorry, web server for RQ failed to respond correctly"
+  exit 1
+fi
+
+egrep "successqueue created" _install_test_nop.txt > /dev/null
+if [ "$?" -ne "0" ]; then
+  echo "Sorry, system didn't create test_nop queue"
+  exit 1
+fi
+
 rm _home.txt
 rm _install.txt
 rm _install_test.txt
 rm _install_test_coalesce.txt
 rm _install_test_run.txt
+rm _install_test_nop.txt
 
 echo "ALL DONE SUCCESSFULLY"
